Run pil source tests in the container as opposed to host
